src/main/java/com/system/juan/miguel/notification/controller/NotificationController.java
package com.system.juan.miguel.notification.controller;

import com.system.juan.miguel.notification.entity.NotificationLog;
import com.system.juan.miguel.notification.service.NotificationService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Map;

@RestController
@RequestMapping("/api/notifications")
@CrossOrigin(origins = "http://localhost:3000")
public class NotificationController {

    @Autowired
    private NotificationService notificationService;

    @PostMapping("/send")
    public ResponseEntity<?> sendNotification(@RequestBody Map<String, String> payload) {
        notificationService.sendNotification(payload.get("category"), payload.get("message"));
        return ResponseEntity.ok().build();
    }

    @GetMapping("/logs")
    public ResponseEntity<List<NotificationLog>> getLogs() {
        return ResponseEntity.ok(notificationService.getLogs());
    }
}